clear out test cases
