+++ pyproject.toml
@@ -21,7 +21,7 @@ classifiers = [
 requires-python = ">=3.8,<3.11"
 dependencies = [
     "anywidget",
-    "cev-metrics"
+    "cev-metrics",
     "ipywidgets>=8.0.0",
     "jinja2",
     "jupyter-scatter>=0.9.0",
